+++ client/sub_client_output.c
@@ -474,7 +474,7 @@ static void formatted_print_percent(const struct mosq_config *lcfg, const struct
 
 		case 'E':
 			if(mosquitto_property_read_int32(properties, MQTT_PROP_MESSAGE_EXPIRY_INTERVAL, &i32value, false)){
-				formatted_print_int(i32value, align, pad, field_width);
+				formatted_print_int((int)i32value, align, pad, field_width);
 			}else{
 				formatted_print_blank(pad, field_width);
 			}
@@ -509,7 +509,7 @@ static void formatted_print_percent(const struct mosq_config *lcfg, const struct
 					return;
 				}
 			}
-			if(json_print(message, properties, ti, ns, true, lcfg->pretty) != MOSQ_ERR_SUCCESS){
+			if(json_print(message, properties, ti, (int)ns, true, lcfg->pretty) != MOSQ_ERR_SUCCESS){
 				err_printf(lcfg, "Error: Out of memory.\n");
 				return;
 			}
@@ -522,7 +522,7 @@ static void formatted_print_percent(const struct mosq_config *lcfg, const struct
 					return;
 				}
 			}
-			rc = json_print(message, properties, ti, ns, false, lcfg->pretty);
+			rc = json_print(message, properties, ti, (int)ns, false, lcfg->pretty);
 			if(rc == MOSQ_ERR_NOMEM){
 				err_printf(lcfg, "Error: Out of memory.\n");
 				return;
@@ -585,7 +585,7 @@ static void formatted_print_percent(const struct mosq_config *lcfg, const struct
 
 		case 'S':
 			if(mosquitto_property_read_varint(properties, MQTT_PROP_SUBSCRIPTION_IDENTIFIER, &i32value, false)){
-				formatted_print_int(i32value, align, pad, field_width);
+				formatted_print_int((int)i32value, align, pad, field_width);
 			}else{
 				formatted_print_blank(pad, field_width);
 			}
@@ -620,7 +620,7 @@ static void formatted_print_percent(const struct mosq_config *lcfg, const struct
 
 static void formatted_print(const struct mosq_config *lcfg, const struct mosquitto_message *message, const mosquitto_property *properties)
 {
-	int len;
+	size_t len;
 	int i;
 	struct tm *ti = NULL;
 	long ns;
@@ -759,7 +759,7 @@ void rand_init(void)
 	long ns;
 
 	if(!get_time(&ti, &ns)){
-		srandom(ns);
+		srandom((unsigned int)ns);
 	}
 #endif
 }
